Fix rendering process card loop
